requirements.txt
amqp==2.6.1
APScheduler==3.6.3
asgiref==3.2.10
billiard==3.6.4.0
cachetools==4.2.2
celery==4.4.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==2.2.28
djangorestframework==3.12.4
flower==0.9.7
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.0.4
humanize==4.4.0
idna==3.3
kombu==4.6.11
Pillow==9.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.32
psycopg2-binary==2.8.6
pycparser==2.21
python-dotenv==0.19.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
telegram==0.0.1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.8
vine==1.3.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.5.2
